Use the Settings API
Modified the class to use the new, built in settings api.  This takes care of a
lot of the issues like making sure the user can save the settings, making sure
nonces get validated etc.

Also cleaned up some code and modified the loop of post types to exclude
attachments, which never have a "publish" status and won't show up on the front
end without further modification.

Finally, the text strings are now translatable -- in case you want to
internationalize this plugin.